Deleted unnecessary modules
