+++ pytools/__init__.py
@@ -1,5 +1,4 @@
 from . import tabletools
 from . import timetools 
 from . import systemtools 
-from . import filetools 
-from . import widgets
+from . import filetools
