+++ pytools/filetools/_filetools.py
@@ -1,9 +1,29 @@
 
 import hashlib
-import os
-from typing import List
-from pathlib import Path
 
+from pathlib import Path
+import os
+import psutil
+
+def memoryUsage(show = True, units = 'MB'):
+	""" Gets the current memory usage
+		Returns
+		----------
+			if show is False
+			memory: int
+				The total number of bytes being used by the current process
+	"""
+
+	process = psutil.Process(os.getpid())
+	usage = process.memory_info().rss
+	if show:
+		if units == 'MB':
+			value = usage / 1024**2
+		else:
+			value = usage
+		print("Current memory usage: {0:.2f}{1}".format(value, units), flush = True)
+	else:
+		return usage
 
 def checkDir(path:Path, full = False):
 	""" Creates a folder if it doesn't already exist.
