--- pytools/systemtools/__init__.py
@@ -1,2 +1,0 @@
-from ._terminal import Terminal
-from ._system_methods import *
