--- pytools/systemtools/_system_methods.py
@@ -1,25 +1,0 @@
-import os
-import psutil
-
-def memoryUsage(show = True, units = 'MB'):
-	""" Gets the current memory usage 
-		Returns
-		----------
-			if show is False
-			memory: int
-				The total number of bytes being used by the current process
-	"""
-
-	process = psutil.Process(os.getpid())
-	usage = process.memory_info().rss
-	if show:
-		if units == 'MB':
-			value = usage / 1024**2
-		else:
-			value = usage
-		print("Current memory usage: {0:.2f}{1}".format(value, units), flush = True)
-	else:
-		return usage
-
-if __name__ == "__main__":
-	memoryUsage()
